# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dokoopy 1.0.0",
    "description": "API for charity landing page of 57th brigade.\n\nThis Server based on the OpenAPI 3.0 specification\n\nSome useful links:\n- [The Dokoopy-backend repository](https://github.com/baza-trainee/dokoopy-backend)",
    "contact": {
      "name": "API developers",
      "email": ""
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5001/api",
      "description": "Development Server"
    },
    {
      "url": "https://dokoopy.onrender.com/api",
      "description": "Production Server"
    },
    {
      "url": "https://dokoopy-pr-19.onrender.com/api",
      "description": "Test Server for feature get-main-content"
    }
  ],
  "consumes": ["application/json", "multipart/form-data", "text/html"],
  "produces": ["application/json", "multipart/form-data", "text/html"],
...
  "paths": {
...
    "/auth/admin/forgot-password": {
      "post": {
        "summary": "Send forgot-password email to the user",
        "tags": [
          "Auth"
        ],
        "description": "Send the email with link for resetting password",
        "operationId": "forgotPassword",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "String",
                  "example": "Reset password email sent"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/auth/admin/reset-password/{resetToken}": {
      "post": {
        "summary": "Reset password",
        "tags": [
          "Auth"
        ],
        "description": "Write new password to the Database",
        "operationId": "resetPassword",
        "parameters": [
          {
            "name": "resetToken",
            "in": "params",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Backend-generated unique resetToken"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "String",
                  "example": "Reset password is succesful"
                }
              }
            }
          },
          "403": {
            "description": "Reset token is expired"
          }
        }
      }
    },
...
    "/contacts/admin/{id}": {
      "patch": {
        "summary": "Update contacts info",
        "tags": [
          "Contacts"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Update contacts info in the database",
        "operationId": "updateContact",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "dokoopy@gmail.com"
                  },
                  "telegram": {
                    "type": "string",
                    "example": "@dokoopy"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateContactResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          }
        }
      },
      "get": {
        "summary": "Get all reports for the admin panel",
        "tags": [
          "Contacts"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Get all contacts for the admin panel",
        "operationId": "getAllContacts",
        "responses": {
          "200": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GetAllContactsResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          }
        }
      }
    },
    "/contacts": {
      "get": {
        "summary": "Get all reports for the main page",
        "tags": [
          "Contacts"
        ],
        "description": "Get all contacts for the main page",
        "operationId": "getAllContacts",
        "responses": {
          "200": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GetAllContactsResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          }
        }
      }
    },
...
    "/hero/admin": {
      "post": {
        "summary": "Add new hero",
        "tags": [
          "Heroes"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Add new hero",
        "operationId": "addHero",
        "requestBody": {
            "required": true,
            "content": {
              "multipart/form-data": {
                "schema": {
                  "$ref": "#/components/schemas/AddHeroRequest"
                }
              }
            }
        },
        "responses": {
          "201": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AddHeroResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          },
          "500": {
            "description": "File too large"
          }
        }
      },
      "get": {
        "summary": "Get all heroes for the admin panel",
        "tags": [
          "Heroes"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Get all heroes for the admin panel",
        "operationId": "getAllHeroes",
        "responses": {
          "200": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GetAllHeroesResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          }
        }
      }
    },
    "/hero/admin/{id}": {
      "patch": {
        "summary": "Update hero info",
        "tags": [
          "Heroes"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Update hero information in the database",
        "operationId": "updateHero",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "example": "6490564f8f85c705511c5de2",
            "schema": {
              "type": "string",
              "description": "Backend-generated unique identifier"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHeroRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddHeroResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          },
          "500": {
            "description": "File too large"
          }
        }
      },
      "delete": {
          "tags": [
            "Heroes"
          ],
          "security": [
            {
              "Bearer": []
            }
          ],
          "summary": "Delete the hero",
          "description": "Delete the hero from the database",
          "operationId": "deleteHero",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of hero that needs to be deleted",
              "required": true,
              "schema": {
                "type": "string",
                "description": "Delete the hero with the specified ID"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Hero deleted"
            },
            "401": {
              "description": "Not authorized"
            }
          }
      }
    },
    "/hero": {
      "get": {
        "summary": "Get all heroes for the main page",
        "tags": [
          "Heroes"
        ],
        "description": "Get all heroes for the main page",
        "operationId": "getAllHeroes",
        "responses": {
          "200": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GetAllHeroesResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          }
        }
      }
    },
    "/bank/admin/{id}": {
      "patch": {
        "summary": "Update jar link",
        "tags": [
          "Bank"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Update jar link in the database",
        "operationId": "updateBank",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "link": {
                    "type": "string",
                    "example": "monobank_link"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateBankLinkResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authorized"
          }
        }
      },
      "get": {
        "summary": "Get jar link for the admin panel",
        "tags": [
          "Bank"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Get jar link for the admin panel",
        "operationId": "getAllRequisites",
        "responses": {
          "200": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GetAllRequisitesResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          }
        }
      }
    },
    "/bank": {
      "get": {
        "summary": "Get jar link for the main page",
        "tags": [
          "Bank"
        ],
        "description": "Get jar link for the main page",
        "operationId": "getAllRequisites",
        "responses": {
          "200": {
              "description": "success",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/GetAllRequisitesResponse"
                  }
                }
              }
          },
          "401": {
              "description": "Not authorized"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ForgotPasswordRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "description": "user email",
            "example": "user@gmail.com"
          }
        }
      },
      "ResetPasswordRequest": {
        "type": "object",
        "required": [
          "password"
        ],
        "properties": {
          "password": {
            "type": "string",
            "description": "new password",
            "example": "fhdjhfjshh_123"
          }
        }
      },
...
      "UpdateContactResponse": {
        "type": "object",
            "properties": {
                "_id": {
                  "type": "string",
                  "description": "Backend-generated unique identifier",
                  "example": "6525c81a0337252aedd6124d"
                },
                "email": {
                  "type": "string",
                  "example": "dokoopy@gmail.com"
                },
                "telegram": {
                  "type": "string",
                  "example": "@dokoopy"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2023-10-09T22:37:14.832Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2023-10-09T22:37:14.832Z"
                }
            }
      },
...
      "AddHeroRequest": {
        "type": "object",
        "required": [
          "title",
          "description",
          "description_eng",
          "imageURL"
        ],
        "properties": {
          "title": {
              "type": "string",
              "example": "Hero 1"
          },
          "description": {
              "type": "string",
              "example": "Це опис слайдера"
          },
          "description_eng": {
            "type": "string",
            "example": "This is the super hero description"
        },
          "imageURL": {
              "type": "string",
              "format": "binary"
          }
        }
      },
      "AddHeroResponse": {
          "type": "object",
            "properties": {
              "_id": {
                "type": "string",
                "description": "Backend-generated unique identifier",
                "example": "GGHJlhkj668__hjkhjkhh$%"
              },
              "title": {
                "type": "string",
                "description": "Project title",
                "example": "Hero 1"
              },
              "description": {
                "type": "string",
                "description": "Project description",
                "example": "Це опис слайдера"
              },
              "description_eng": {
                "type": "string",
                "description": "Project description",
                "example": "This is the super hero description"
              },
              "imageURL": {
                  "type": "string",
                  "description": "Project image url",
                  "example": "heroes\\qtc7YIPFu7y55n-hrA0Nl_IMG_7842.JPG"
              },
              "createdAt": {
                "type": "string",
                "example": "2023-05-15T10:51:30.531Z"
              },
              "updatedAt": {
                "type": "string",
                "example": "2023-05-15T10:51:30.531Z"
              }
          }
      },
      "UpdateHeroRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Hero 1"
                    },
                    "description": {
                        "type": "string",
                        "example": "Це опис слайдера"
                    },
                    "description_eng": {
                        "type": "string",
                        "example": "This is the super hero description"
                    },
                    "imageURL": {
                        "type": "string",
                        "format": "binary"
                    }
                }
      },
      "UpdateBankLinkResponse": {
        "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "description": "Backend-generated unique identifier",
              "example": "mono_link"
            },
            "link": {
              "type": "string",
              "example": "mono_link"
            }
          }
      },
...
      "GetAllContactsResponse": {  
        "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "description": "Backend-generated unique identifier",
              "example": "6525c81a0337252aedd6124d"
            },
            "email": {
              "type": "string",
              "example": "dokoopy@gmail.com"
            },
            "telegram": {
              "type": "string",
              "example": "@dokoopy"
            }
        }
      },
      "GetAllRequisitesResponse": {  
        "type": "object",
          "properties": {
            "_id": {
              "type": "string",
              "description": "Backend-generated unique identifier",
              "example": "6525c81a0337252aedd6124d"
            },
            "link": {
              "type": "string",
              "example": "mono_link"
            }
        }
      },
      "GetAllHeroesResponse": {
        "type": "object",
        "properties": {
          "result": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "description": "Backend-generated unique identifier",
                  "example": "GGHJlhkj668__hjkhjkhh$%"
                },
                "title": {
                  "type": "string",
                  "description": "The title of hero",
                  "example": "Hero 1"
                },
                "description": {
                  "type": "string",
                  "description": "The description of hero",
                  "example": "Це перший слайдер"
                },
                "description_eng": {
                  "type": "string",
                  "description": "The description of hero",
                  "example": "This is the first hero slide"
                },
                "imageURL": {
                  "type": "string",
                  "description": "Link on hero's photo",
                  "example": "https://res.cloudinary.com/dwvdhkbkm/image/upload/v1697059077/dokoopy/gr5fchy2vggygzpxnk24.jpg"
                },
                "createdAt": {
                  "type": "string",
                  "example": "2023-09-06T13:06:19.775Z"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "2023-09-06T13:06:19.775Z"
                }
              }
            },
            "description": "List of projects found by request"
          }
        }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}